+++ src/utils/CustomStruct.h
@@ -5,7 +5,7 @@ namespace CustomStruct {
         int widthOffset;
         int heightOffset;
         float size;
-        int opacity;
+        float opacity;
         cocos2d::ccColor3B color;
         std::string separator;
